requirements.txt
openai>=1.0.0
openai-agents>=0.3.0
python-dotenv>=1.0.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
requests>=2.31.0
datasets>=2.14.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0